requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
coverage==5.4
Django==3.1.5
djangorestframework==3.12.2
ics==0.7
idna==2.10
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
TatSu==5.5.0
tqdm==4.56.0
urllib3==1.26.3
pypi==2.1
psycopg2-binary==2.8.6
django-redis==4.12.1
gunicorn==20.0.4
celery==5.0.5